feat(DeleteTerrainById): adiciona validação para garantir que é o usuário que está deletando  o terreno

+++ src/application/usecases/terrain/delete-terrain-by-id.js
@@ -1,4 +1,4 @@
-import { TerrainNotFoundError } from '../../errors'
+import { ForbiddenOperation, TerrainNotFoundError } from '../../errors'
 import { TerrainRepository } from '../../repositories/terrain-repository'
 
 export class DeleteTerrainById {
@@ -6,11 +6,20 @@ export class DeleteTerrainById {
     this.terrainRepository = terrainRepository
   }
 
-  async execute(id) {
+  async execute(id, userIdTryingToDelete) {
     const terrain = await this.terrainRepository.findById(id)
 
     if (!terrain) {
       throw new TerrainNotFoundError()
+    }
+
+    const userIsNotOwner = terrain.userId !== userIdTryingToDelete
+
+    if (userIsNotOwner) {
+      throw new ForbiddenOperation({
+        action:
+          'Para deletar um terreno você deve ser dono dele! Verifique seu acesso e tente novamente.',
+      })
     }
 
     await this.terrainRepository.delete(id)
